+++ changelog.md
@@ -3,6 +3,7 @@
 ## dev
 
 CodeEditor:
+- Ctrl+Click now can open file paths (added `options.onOpenPath`).
 - Not saved dot indicator.
 - Support scroll on click outside Scrollbar thumb.
 - Fixed mousemove error in doc when no tab.